refactor: create hash file
